requirements.txt
pytest==6.2.1
requests==2.25.1
allure-pytest==2.8.31
PyHamcrest==2.0.2
deepdiff==5.2.3

